separate out row panel
