--- src/datascience-ui/data-explorer/data-wrangler/controls/column-operations/DropMissingColumnsSection.tsx
@@ -1,34 +1,0 @@
-import * as React from 'react';
-import {
-    DataWranglerCommands,
-    IDropNaRequest
-} from '../../../../../client/datascience/data-viewing/data-wrangler/types';
-import { buttonStyle } from '../styles';
-
-interface IProps {
-    selectedColumns: string[];
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    submitCommand(data: { command: string; args: any }): void;
-    setColumns(cols: number[]): void;
-}
-
-interface IState {}
-
-export class DropMissingColumnsSection extends React.Component<IProps, IState> {
-    render() {
-        return (
-            <button
-                onClick={() => {
-                    this.props.submitCommand({
-                        command: DataWranglerCommands.DropNa,
-                        args: { targetColumns: this.props.selectedColumns } as IDropNaRequest
-                    });
-                    this.props.setColumns([]);
-                }}
-                style={buttonStyle}
-            >
-                Drop
-            </button>
-        );
-    }
-}
